Create findResultSetNames function
